requirements.txt
softlayer
pyyaml
